+++ example/lib/cards_controller.dart
@@ -2,7 +2,7 @@ import 'package:flutter_control/core.dart';
 
 import 'cards_page.dart';
 
-class CardsController extends BaseController with RouteController {
+class CardsController extends BaseController with RouteController, LocalizationProvider {
   final cards = ListControl<CardModel>();
   final countLabel = StringControl();
 
@@ -19,7 +19,11 @@ class CardsController extends BaseController with RouteController {
     factory.subscribe<CardModel>('remove_card', (card) => removeCard(card));
   }
 
-  void addCard() => newCard('${localize('card_title')} ${_counter++}');
+  void addCard() {
+    newCard('${localize('card_title')} $_counter${localizePlural('number', _counter)}');
+
+    _counter++;
+  }
 
   CardModel newCard(String title) {
     if (title == null || title.isEmpty) {
@@ -46,7 +50,7 @@ class CardsController extends BaseController with RouteController {
   }
 }
 
-class DetailController extends BaseController with RouteController {
+class DetailController extends BaseController with RouteController, LocalizationProvider {
   CardModel _model;
 
   ListControl<CardItemModel> get items => _model.items;
@@ -57,7 +61,7 @@ class DetailController extends BaseController with RouteController {
   void onInit(Map<String, dynamic> args) {
     super.onInit(args);
 
-    _model = ArgProvider.map<CardModel>(args);
+    _model = ArgHandler.map<CardModel>(args);
 
     _model.countLabel.streamTo(title, converter: (input) => '${_model.title} - $input');
   }
@@ -89,7 +93,7 @@ class DetailController extends BaseController with RouteController {
   }
 }
 
-class CardModel extends BaseModel {
+class CardModel extends BaseModel with LocalizationProvider {
   final String title;
   final countLabel = StringControl();
   final progress = DoubleControl();