+++ lib/src/controller/base_model.dart
@@ -1,6 +1,7 @@
 import 'package:flutter_control/core.dart';
 
 /// Lightweight version of Controller. Mainly used for Items in dynamic List or to separate/reuse Logic.
+/// Mixin your model with [LocalizationProvider] to enable localization.
 class BaseModel implements Initializable, Disposable {
   /// returns instance of [ControlFactory] if available.
   /// nullable
@@ -10,26 +11,11 @@ class BaseModel implements Initializable, Disposable {
   /// nullable
   AppControl get control => factory.get(ControlKey.control);
 
-  /// returns instance of [BaseLocalization]
-  BaseLocalization get _localization => factory.get(ControlKey.localization);
-
   /// Default constructor.
   BaseModel();
 
   @override
   void init(Map args) {}
-
-  /// Tries to localize text by given key.
-  /// Localization is part of AppControl or BaseApp Widget.
-  /// Non null.
-  @protected
-  String localize(String key) => _localization?.localize(key) ?? '';
-
-  /// Tries to localize text by given key.
-  /// Localization is part of AppControl or BaseApp Widget.
-  /// Non null.
-  @protected
-  String extractLocalization(Map field) => _localization?.extractLocalization(field) ?? '';
 
   @override
   void dispose() {}