+++ lib/src/util/future_block.dart
@@ -11,6 +11,7 @@ class FutureBlock {
   /// Returns true if last delay is in progress.
   bool get isActive => _timer != null && _timer.isActive;
 
+  /// Default constructor.
   FutureBlock();
 
   /// Starts delay for given [duration]. Given callback can be postponed or canceled.
@@ -50,13 +51,24 @@ class FutureBlock {
     _callback = null;
   }
 
+  /// Runs delayed [Future] with zero [Duration] so [action] will be performed next frame.
   static Future nextFrame(VoidCallback action) => Future.delayed(const Duration(), action);
 }
 
+/// Helps to block part of code for minimum-given time.
+/// Just wrap code with [start] and [finish].
+/// If code runs fast, then finish is awaited for rest of [duration].
+/// If code runs too slowly, then finish is triggered immediately.
 class DelayBlock {
+  /// Delay in milliseconds.
   int _millis;
+
+  /// Timestamp of block start.
   DateTime _start;
 
+  /// Default constructor
+  /// [duration] of delay block.
+  /// [startNow] immediately in constructor.
   DelayBlock(Duration duration, [bool startNow = true]) {
     _millis = duration.inMilliseconds;
     if (startNow) {
@@ -64,10 +76,15 @@ class DelayBlock {
     }
   }
 
+  /// Sets start timestamp.
+  /// Can be called multiple times - timestamp is updated and delay postponed.
   void start() {
     _start = DateTime.now();
   }
 
+  /// awaits delay finish.
+  /// If code runs fast, then finish is awaited for rest of [duration].
+  /// If code runs too slowly, then finish is triggered immediately.
   Future<void> finish() {
     final currentDelay = DateTime.now().difference(_start).inMilliseconds;
     final delay = _millis - currentDelay;